Move config logging from main to res_config
