+++ src/ert/logging/logger.conf
@@ -93,6 +93,9 @@ loggers:
     level: ERROR
     handlers: [eefile]
     propagate: yes
+  ert._c_wrappers.enkf.res_config:
+    level: INFO
+    propagate: yes
 root:
   level: DEBUG
   handlers: [file]
